+++ AdvinstTool/AdvinstToolCleanup.ts
@@ -3,16 +3,19 @@ import * as path from 'path';
 
 async function run() {
   try {
-    if (taskLib.osType() != 'Windows_NT')
-      throw new Error('Only Windows systems are supported.');
+    taskLib.setResourcePath(path.join(__dirname, "task.json"));
+    if (taskLib.osType() != 'Windows_NT') {
+      console.log(taskLib.loc("UnsupportedOS"));
+      return;
+    }
 
     const performCleanup: string = taskLib.getVariable('advinst.cleanup');
     taskLib.debug('advinst.cleanup = ' + performCleanup);
     if (!performCleanup) {
       return;
     }
-    
-    const licensePath : string = path.join(taskLib.getVariable('ProgramData'), 'Caphyon\\Advanced Installer\\license80.dat');
+
+    const licensePath: string = path.join(taskLib.getVariable('ProgramData'), 'Caphyon\\Advanced Installer\\license80.dat');
     if (taskLib.exist(licensePath)) {
       taskLib.rmRF(licensePath);
     }